requirements.txt
aiogram
asyncpg
python-dotenv
matplotlib
seaborn
requests
pdfkit
jinja2
scikit-learn
